Add tests for Lion class and handle exceptions
